src/infrastructure/error/error_handling_system.py
"""Error handling system implementation."""

import logging
import traceback
from typing import Any, Dict, Optional


class ErrorHandlingSystem:
    """Central error handling system."""

    def __init__(self) -> None:
        """Initialize error handling system."""
        self.logger = logging.getLogger("error_handler")
        self.setup_logger()

    def setup_logger(self) -> None:
        """Set up logging configuration."""
        # TODO: Implement logger setup
        pass

    def handle_error(
        self, error: Exception, context: Optional[Dict[str, Any]] = None
    ) -> None:
        """Handle an error with context."""
        # TODO: Implement error handling
        pass

    def log_error(self, error: Exception, stack_trace: str) -> None:
        """Log an error with its stack trace."""
        # TODO: Implement error logging
        pass

    def recover_from_error(self, error: Exception) -> bool:
        """Attempt to recover from an error."""
        # TODO: Implement error recovery
        return False
