tests/unit/test_error_handling_system.py
"""Tests for error handling system."""

import pytest

from src.infrastructure.error.error_handling_system import ErrorHandlingSystem


def test_error_handling_system_init():
    """Test error handling system initialization."""
    system = ErrorHandlingSystem()
    assert system is not None


def test_handle_error():
    """Test error handling."""
    system = ErrorHandlingSystem()
    error = ValueError("Test error")
    system.handle_error(error)
    # TODO: Add assertions


def test_error_recovery():
    """Test error recovery."""
    system = ErrorHandlingSystem()
    error = ValueError("Test error")
    result = system.recover_from_error(error)
    assert isinstance(result, bool)
